[development/globals] refactor of globals usage in the engine. the rule is not that you have an init func for your global, an instance func and a shutdown func. This keeps it simple and predictable when you will initialize, access and destroy the global. As a result, the engine has changed a bit and common paths are now part of the engine global

+++ source/1_engine/rex_engine/include/public/rex_engine/memory/blob.h
@@ -82,7 +82,7 @@ namespace rex
       {
         REX_ASSERT_X(m_data.count() % sizeof(T) == 0, "You can't release a blob if you can't fit all the data in the resulting array");
 
-        s32 count = m_data.count();
+        s32 count = static_cast<s32>(m_data.count());
         T* data_ptr = reinterpret_cast<T*>(m_data.release());
         return rsl::unique_array<T>(data_ptr, count / sizeof(T));
       }
